EES-3829 - general cleanup of python test code
