+++ Pipfile
@@ -20,6 +20,9 @@ slack-sdk = "~=3.18.3"
 lxml = "~=4.9.1"
 azure-storage-blob = "~=12.9.0"
 autopep8 = "~=1.7.0"
-
+scikit-image = "~=0.19.2"
+imutils = "~=0.5.4"
+opencv-python = "~=4.5.5.64"
+numpy = "~=1.22.0"
 [requires]
 python_version = "3.10.5"
