+++ tests/robot-tests/run_tests.py
@@ -7,24 +7,18 @@ Run 'python run_tests.py -h' to see argument options
 """
 
 import argparse
-import cProfile
 import datetime
-import json
 import os
-import pstats
 import shutil
 from pathlib import Path
-import requests
 from dotenv import load_dotenv
 from pabot.pabot import main as pabot_run_cli
 from robot import rebot_cli as robot_rebot_cli
 from robot import run_cli as robot_run_cli
 from scripts.get_webdriver import get_webdriver
-import scripts.keyword_profile as kp
-from tests.libs.setup_auth_variables import setup_auth_variables
 from tests.libs.slack import send_slack_report
 from tests.libs.create_emulator_release_files import ReleaseFilesGenerator
-
+from tests.libs.admin_api import setup_authentication, create_test_theme, create_test_topic, get_test_themes, delete_test_topic
 
 current_dir = Path(__file__).absolute().parent
 os.chdir(current_dir)
@@ -45,21 +39,14 @@ else:
 parser = argparse.ArgumentParser(
     prog="pipenv run python run_tests.py",
     description="Use this script to run the UI tests, locally or as part of the CI pipeline, against the environment of your choosing")
-parser.add_argument(
-    "-b",
-    "--browser",
-    dest="browser",
-    default="chrome",
-    choices=[
-        "chrome",
-        "firefox",
-        "ie"],
-    help="name of the browser you wish to run the tests with (NOTE: Only chromedriver is automatically installed!)")
+
 parser.add_argument("-i", "--interp",
                     dest="interp",
                     default="pabot",
                     choices=["pabot", "robot"],
-                    help="interpreter to use to run the tests")
+                    help="interpreter to use to run the tests",
+                    required=True,
+                    )
 parser.add_argument("--processes",
                     dest="processes",
                     help="how many processes should be used when using the pabot interpreter")
@@ -67,12 +54,14 @@ parser.add_argument("-e", "--env",
                     dest="env",
                     default="test",
                     choices=["local", "dev", "test", "preprod", "prod", "ci"],
-                    help="the environment to run the tests against")
+                    help="the environment to run the tests against",
+                    required=True)
 parser.add_argument("-f", "--file",
                     dest="tests",
                     metavar="{file/dir}",
                     default="tests/",
-                    help="test suite or folder of tests suites you wish to run")
+                    help="test suite or folder of tests suites you wish to run",
+                    required=True)
 parser.add_argument("-t", "--tags",
                     dest="tags",
                     nargs="?",
@@ -82,10 +71,6 @@ parser.add_argument("-v", "--visual",
                     dest="visual",
                     action="store_true",
                     help="display browser window that the tests run in")
-parser.add_argument("-p", "--profile",
-                    dest="profile",
-                    action="store_true",
-                    help="output profiling information")
 parser.add_argument("--ci",
                     dest="ci",
                     action="store_true",
@@ -112,7 +97,7 @@ parser.add_argument("--print-keywords",
                     help="choose to print out keywords as they are started")
 parser.add_argument("--enable-slack",
                     dest="enable_slack",
-                    action='store_true'
+                    action='store_true',
                     )
 parser.add_argument("--prompt-to-continue",
                     dest="prompt_to_continue",
@@ -135,7 +120,7 @@ environment variables, and instead must be passed as an argument to this script.
 parser.add_argument("--slack-webhook-url",
                     dest="slack_webhook_url",
                     default=None,
-                    help="URL for Slack webhook")
+                    help="URL for Slack webhook (responsible for sending test reports to slack)")
 parser.add_argument("--admin-pass",
                     dest="admin_pass",
                     default=None,
@@ -146,11 +131,7 @@ parser.add_argument("--analyst-pass",
                     help="manually specify the analyst password")
 args = parser.parse_args()
 
-
-if args.custom_env:
-    load_dotenv(args.custom_env)
-else:
-    load_dotenv('.env.' + args.env)
+load_dotenv(args.custom_env) if args.custom_env else load_dotenv('.env.' + args.env)
 
 assert os.getenv('TIMEOUT') is not None
 assert os.getenv('IMPLICIT_WAIT') is not None
@@ -165,7 +146,6 @@ assert os.getenv('WAIT_SMALL') is not None
 assert os.getenv('FAIL_TEST_SUITES_FAST') is not None
 assert os.getenv('IDENTITY_PROVIDER') is not None
 assert os.getenv('WAIT_MEMORY_CACHE_EXPIRY') is not None
-
 
 if args.slack_webhook_url:
     os.environ['SLACK_WEBHOOK_URL'] = args.slack_webhook_url
@@ -214,124 +194,10 @@ if args.ci:
     robotArgs += ['--removekeywords',
                   'name:common.user goes to url']  # To hide basic auth credentials
 
-# seed Azure storage emulator release files
-if (args.env == 'local'):
-    generator = ReleaseFilesGenerator()
-    generator.create_public_release_files()
-    generator.create_private_release_files()
-
-
-def admin_request(method, endpoint, body=None):
-    assert method and endpoint
-    assert os.getenv('ADMIN_URL') is not None
-    assert os.getenv('IDENTITY_LOCAL_STORAGE_ADMIN') is not None
-
-    if method == 'POST':
-        assert body is not None, 'POST requests require a body'
-
-    requests.sessions.HTTPAdapter(
-        pool_connections=50,
-        pool_maxsize=50,
-        max_retries=3
-    )
-    session = requests.Session()
-
-    # To prevent InsecureRequestWarning
-    requests.packages.urllib3.disable_warnings()
-
-    jwt_token = json.loads(os.getenv('IDENTITY_LOCAL_STORAGE_ADMIN'))['access_token']
-    headers = {
-        'Content-Type': 'application/json',
-        'Authorization': f'Bearer {jwt_token}',
-    }
-    response = session.request(
-        method,
-        url=f'{os.getenv("ADMIN_URL")}{endpoint}',
-        headers=headers,
-        stream=True,
-        json=body,
-        verify=False
-    )
-
-    if response.status_code in {401, 403}:
-        print('Attempting re-authentication...', flush=True)
-
-        # Delete identify files and re-attempt to fetch them
-        setup_authentication(clear_existing=True)
-        jwt_token = json.loads(os.environ['IDENTITY_LOCAL_STORAGE_ADMIN'])['access_token']
-        response = session.request(
-            method,
-            url=f'{os.getenv("ADMIN_URL")}{endpoint}',
-            headers={
-                'Content-Type': 'application/json',
-                'Authorization': f'Bearer {jwt_token}',
-            },
-            stream=True,
-            json=body,
-            verify=False
-        )
-
-        assert response.status_code not in {401, 403}, \
-            'Failed to reauthenticate.'
-
-    assert response.status_code < 300, f'Admin request responded with {response.status_code} and {response.text}'
-    return response
-
-
-def get_test_themes():
-    return admin_request('GET', '/api/themes')
-
-
-def create_test_theme():
-    return admin_request('POST', '/api/themes', {
-        'title': 'Test theme',
-        'summary': 'Test theme summary'
-    })
-
-
-def create_test_topic():
-    assert os.getenv('TEST_THEME_ID') is not None
-
-    topic_name = f'UI test topic {os.getenv("RUN_IDENTIFIER")}'
-    resp = admin_request('POST', '/api/topics', {
-        'title': topic_name,
-        'themeId': os.getenv('TEST_THEME_ID')
-    })
-
-    os.environ['TEST_TOPIC_NAME'] = topic_name
-    os.environ['TEST_TOPIC_ID'] = resp.json()['id']
-
-
-def delete_test_topic():
-    if os.getenv('TEST_TOPIC_ID') is not None:
-        admin_request('DELETE', f'/api/topics/{os.getenv("TEST_TOPIC_ID")}')
-
-
-def setup_authentication(clear_existing=False):
-    # Don't need BAU user if running general_public tests
-    if "general_public" not in args.tests:
-        setup_auth_variables(
-            user='ADMIN',
-            email=os.getenv('ADMIN_EMAIL'),
-            password=os.getenv('ADMIN_PASSWORD'),
-            clear_existing=clear_existing,
-            identity_provider=os.getenv('IDENTITY_PROVIDER')
-        )
-
-    # Don't need analyst user if running admin/bau or admin_and_public/bau tests
-    if f"{os.sep}bau" not in args.tests:
-        setup_auth_variables(
-            user='ANALYST',
-            email=os.getenv('ANALYST_EMAIL'),
-            password=os.getenv('ANALYST_PASSWORD'),
-            clear_existing=clear_existing,
-            identity_provider=os.getenv('IDENTITY_PROVIDER')
-        )
-
 
 # Auth not required with general_public tests
 if args.tests and "general_public" not in args.tests:
-    setup_authentication()
+    setup_authentication(args)
 
     # NOTE(mark): Tests that alter data only occur on local and dev environments
     if args.env in ['local', 'dev']:
@@ -357,16 +223,25 @@ if args.tests and "general_public" not in args.tests:
 
         create_test_topic()
 
+
 if args.env == 'local':
     robotArgs += ['--include', 'Local']
     robotArgs += ['--exclude', 'NotAgainstLocal']
+
+    # seed Azure storage emulator release files
+    generator = ReleaseFilesGenerator()
+    generator.create_public_release_files()
+    generator.create_private_release_files()
+
 if args.env == 'dev':
     robotArgs += ['--include', 'Dev']
     robotArgs += ['--exclude', 'NotAgainstDev']
+
 if args.env == 'test':
     robotArgs += ['--include', 'Test',
                   '--exclude', 'NotAgainstTest',
                   '--exclude', 'AltersData']
+
 if args.env == 'preprod':
     robotArgs += ['--include', 'Preprod',
                   '--exclude', 'AltersData',
@@ -391,7 +266,7 @@ if os.getenv('FAIL_TEST_SUITES_FAST'):
 if args.prompt_to_continue:
     robotArgs += ["-v", "prompt_to_continue_on_failure:1"]
 
-robotArgs += ["-v", "browser:" + args.browser]
+robotArgs += ["-v", "browser:" + 'chrome']
 robotArgs += [args.tests]
 
 # Remove any existing test results if running from scratch
@@ -401,48 +276,14 @@ if not args.rerun_failed_tests and not args.rerun_failed_suites and Path('test-r
 try:
     # Run tests
     if args.interp == "robot":
-        if args.profile:
-            # Python profiling
-            cProfile.run('robot_run_cli(robotArgs)', 'profile-data')
-            stream = open('test-results/python-profiling-results.log', 'w', encoding='utf-8')
-            p = pstats.Stats('profile-data', stream=stream)
-            p.sort_stats('time')
-            # p.sort_stats('cumulative')
-            p.print_stats()
-            os.remove('profile-data')
-
-            # Keyword profiling
-            kp.run_keyword_profile(f'test-results/{output_file}',
-                                   printresults=False,
-                                   writepath=f'test-results/keyword-profiling-results.log')
-            print(
-                f'Keyword profiling results saved to {os.getcwd()}/test-results/keyword-profiling-results.log',
-                flush=True)
-        else:
-            robot_run_cli(robotArgs)
+        robot_run_cli(robotArgs)
     elif args.interp == "pabot":
         if args.processes:
             robotArgs = ["--processes", int(args.processes)] + robotArgs
-
-        if args.profile:
-            # Python profiling
-            cProfile.run('pabot_run_cli(robotArgs)', 'profile-data')
-            stream = open(f'test-results{os.sep}python-profiling-results.log', 'w', encoding='utf-8', errors='ignore')
-            p = pstats.Stats('profile-data', stream=stream)
-            p.sort_stats('time')
-            # p.sort_stats('cumulative')
-            p.print_stats()
-            os.remove('profile-data')
-
-            # Keyword profiling
-            kp.run_keyword_profile(f'test-results{os.sep}{output_file}',
-                                   printresults=False,
-                                   writepath=f'test-results/keyword-profiling-results.log')
-            print(
-                f'Keyword profiling results saved to {os.getcwd()}/test-results/keyword-profiling-results.log',
-                flush=True)
-        else:
             pabot_run_cli(robotArgs)
+    else:
+        raise AttributeError(f"Interpreter: '{args.interp}' for running tests not recognized")
+
 
 finally:
     if not args.disable_teardown:
