+++ tests/robot-tests/scripts/get_auth_tokens.py
@@ -1,5 +1,4 @@
 import time
-import os
 import argparse
 import json
 import traceback
