+++ tests/robot-tests/scripts/get_webdriver.py
@@ -5,12 +5,13 @@ import os
 
 
 def get_webdriver(version: str) -> None:
-    chromedriver_filename = 'chromedriver.exe' if platform.system() == "Windows" else 'chromedriver'
+    filename = 'chromedriver.exe' if platform.system() == "Windows" else 'chromedriver'
     pyderman.install(file_directory='./webdriver/',
-                     filename=chromedriver_filename,
+                     filename=filename,
                      verbose=True,
                      chmod=True,
                      overwrite=False,
+                     return_info=True,
                      version=version)
 
     os.environ["PATH"] += os.pathsep + str(Path('./webdriver').absolute())
