--- tests/robot-tests/scripts/visual-testing/Pipfile
@@ -1,15 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-scikit-image = "~=0.19.2"
-imutils = "~=0.5.4"
-opencv-python = "~=4.5.5.64"
-numpy = "~=1.22.0"
-
-[dev-packages]
-
-[requires]
-python_version = "3.10"
