+++ tests/robot-tests/tests/libs/alerts.py
@@ -4,8 +4,7 @@ from slack_sdk.webhook import WebhookClient
 
 def send_slack_alert(err_msg):
     url = os.getenv('SLACK_WEBHOOK_URL')
-    assert url is not None or not "", 'SLACK_WEBHOOK_URL env variable needs to be set'
+    assert url, 'SLACK_WEBHOOK_URL env variable needs to be set'
     slack_webhook = WebhookClient(url)
     response = slack_webhook.send(text=err_msg)
-    assert response.status_code == 200
-    assert response.body == "ok"
+    assert response.status_code == 200 and response.body == 'ok'
